Receptors added back to Debug Render system
