+++ packages/engine/src/ecs/classes/EngineEvents.ts
@@ -38,9 +38,6 @@ const EVENTS = {
 
   CONNECT: 'NETWORK_CONNECT' as const,
 
-  AVATAR_DEBUG: 'CORE_AVATAR_DEBUG' as const,
-  PHYSICS_DEBUG: 'CORE_PHYSICS_DEBUG' as const,
-
   START_SUSPENDED_CONTEXTS: 'POSITIONAL_AUDIO_START_SUSPENDED_CONTEXTS' as const,
   SUSPEND_POSITIONAL_AUDIO: 'POSITIONAL_AUDIO_SUSPEND_POSITIONAL_AUDIO' as const,
